+++ server.js
@@ -18,7 +18,7 @@ if (appData.isDev) {
 }
 
 const { cymreigio } = require('./modules/urls');
-const { getSectionsForNavigation } = require('./controllers/route-helpers');
+const { getSectionsForNavigation } = require('./controllers/helpers/route-helpers');
 const { proxyPassthrough, postToLegacyForm } = require('./modules/legacy');
 const { renderError, renderNotFound, renderUnauthorised } = require('./controllers/http-errors');
 const { SENTRY_DSN } = require('./modules/secrets');
